migrated functions to polling-data file
